Move flags to their own package
To resolve a cyclic dependency issue
