+++ cli/app.go
@@ -15,14 +15,11 @@ import (
 
 var (
 	app         *cli.App
-	globalFlags []cli.Flag
 	config      *cfg.Configuration
 	// ConfigDefaultName is the fallback file name of the YAML file containing the default template values.
 	ConfigDefaultName = "config_defaults.yml"
 	// GrepoSyncFileName is the default file name of the YAML file containing the main settings.
 	GrepoSyncFileName = "greposync.yml"
-
-	logLevelFlagName    = "log-level"
 )
 
 func CreateCLI(version, commit, date string) {
@@ -31,29 +28,6 @@ func CreateCLI(version, commit, date string) {
 	printer.CheckIfError(err)
 
 	config = cfg.NewDefaultConfig()
-	globalFlags = []cli.Flag{
-		&cli.BoolFlag{
-			Name:    "verbose",
-			Aliases: []string{"v"},
-			Usage:   fmt.Sprintf("Shorthand for --%s=debug", logLevelFlagName),
-		},
-		&cli.StringFlag{
-			Name:  logLevelFlagName,
-			Usage: "Log level. Allowed values are [debug, info, warn, error].",
-			Value: config.Log.Level,
-		},
-		&cli.PathFlag{
-			Name:  projectRootFlagName,
-			Usage: "Local directory path where git clones repositories into.",
-			Value: config.Project.RootDir,
-		},
-		&cli.IntFlag{
-			Name:    ProjectJobsFlagName,
-			Usage:   "Jobs is the number of parallel jobs to run. 1 basically means that jobs are run in sequence.",
-			Aliases: []string{"j"},
-			Value:   1,
-		},
-	}
 	app = &cli.App{
 		Name:                 "greposync",
 		Usage:                "git-repo-sync: Shameless reimplementation of ModuleSync in Go",
@@ -75,19 +49,4 @@ func CreateCLI(version, commit, date string) {
 func Run() {
 	err := app.Run(os.Args)
 	printer.CheckIfError(err)
-}
-
-func CombineWithGlobalFlags(flags ...cli.Flag) []cli.Flag {
-	for _, flag := range flags {
-		globalFlags = append(globalFlags, flag)
-	}
-	return globalFlags
-}
-
-func ValidateGlobalFlags(ctx *cli.Context) error {
-	if ctx.Bool("verbose") {
-		config.Log.Level = "debug"
-		printer.DefaultLevel = printer.LevelDebug
-	}
-	return nil
 }
