+++ cli/update/validate.go
@@ -6,29 +6,29 @@ import (
 	"regexp"
 
 	"github.com/ccremer/greposync/cfg"
-	app "github.com/ccremer/greposync/cli"
+	"github.com/ccremer/greposync/cli/flags"
 	"github.com/ccremer/greposync/printer"
 	"github.com/urfave/cli/v2"
 )
 
 func (c *Command) validateUpdateCommand(ctx *cli.Context) error {
-	if err := cfg.ParseConfig(app.GrepoSyncFileName, c.cfg, ctx); err != nil {
+	if err := cfg.ParseConfig(c.cfg.Project.MainConfigFileName, c.cfg, ctx); err != nil {
 		return err
 	}
 
-	if err := app.ValidateGlobalFlags(ctx); err != nil {
+	if err := flags.ValidateGlobalFlags(ctx, c.cfg); err != nil {
 		return err
 	}
 
 	if _, err := regexp.Compile(c.cfg.Project.Include); err != nil {
-		return fmt.Errorf("invalid flag --%s: %v", app.ProjectIncludeFlagName, err)
+		return fmt.Errorf("invalid flag --%s: %v", flags.ProjectIncludeFlagName, err)
 	}
 	if _, err := regexp.Compile(c.cfg.Project.Exclude); err != nil {
-		return fmt.Errorf("invalid flag --%s: %v", app.ProjectExcludeFlagName, err)
+		return fmt.Errorf("invalid flag --%s: %v", flags.ProjectExcludeFlagName, err)
 	}
 
-	if jobs := c.cfg.Project.Jobs; jobs > app.JobsMaximumCount || jobs < app.JobsMinimumCount {
-		return fmt.Errorf("--%s is required to be between %d and %d", app.ProjectJobsFlagName, app.JobsMinimumCount, app.JobsMaximumCount)
+	if jobs := c.cfg.Project.Jobs; jobs > flags.JobsMaximumCount || jobs < flags.JobsMinimumCount {
+		return fmt.Errorf("--%s is required to be between %d and %d", flags.ProjectJobsFlagName, flags.JobsMinimumCount, flags.JobsMaximumCount)
 	}
 
 	if ctx.IsSet(dryRunFlagName) {
